Short Circuits for PyInstaller
Until recently PyInstaller was using a rather old version of pefile,
with some custom modifications to skip processing that wasn't needed in
that application and some changes for PY3 support.  PyInstaller has now
transitioned to using pefile from PyPi and would like to have the
shortcut flags adopted in the official repo.