Fix bug resulting in not parsing all OPTIONAL_HEADER DataDirectory entries
In the format for the PE optional header, the DataDirectory list has a fixed size of 16 address/size entries.
Previously, an incorrect assumption was made that the list of entries is dynamic and given by OPTIONAL_HEADER.NumberOfRvaAndSizes.
This resulted in only parsing the first entries in the list, which might not be the entries that are not empty.
The most extreme case of this was NumberOfRvaAndSizes being 6, and the first 6 data directories were all empty, so none of the directories with actual data were parsed.